modify the auth

+++ src/main/java/com/bracket/BracketAction.java
@@ -4,7 +4,9 @@ import java.util.List;
 
 import org.apache.log4j.Logger;
 
-import com.Constrants;
+import com.Authority;
+import com.Modules;
+import com.Operation;
 import com.ScheduledAction;
 import com.document.Document;
 import com.log.Log;
@@ -42,7 +44,7 @@ public class BracketAction extends ScheduledAction {
 
 	@Override
 	public String getActionModule() {
-		return Constrants.s_bracket_model;
+		return Modules.s_bracket_model;
 	}
 
 	public Bracket getBracket() {
@@ -77,9 +79,13 @@ public class BracketAction extends ScheduledAction {
 	}
 
 	public String bracketAddSubmit() {
+		Authority auth = checkAuthority(buildResource(Modules.s_bracket_model, Operation.s_operation_add));
+		if (auth != null) {
+			return auth.getName();
+		}
 		try {
 			if (m_uploadFile.getFile() != null) {
-				int documentId = m_documentService.insertDocument(Constrants.s_contactChannel_model, m_uploadFile);
+				int documentId = m_documentService.insertDocument(Modules.s_bracket_model, m_uploadFile);
 				m_bracket.setDocumentId(documentId);
 			}
 			m_schedule.setType(getActionModule());
@@ -88,7 +94,7 @@ public class BracketAction extends ScheduledAction {
 
 			int id = m_bracketService.insertBracket(m_bracket);
 			if (id > 0) {
-				Log log = createLog(Constrants.s_bracket_model, Constrants.s_operation_add, m_bracket);
+				Log log = createLog(Modules.s_bracket_model, Operation.s_operation_add, m_bracket);
 
 				m_logService.insertLog(log);
 				return SUCCESS;
@@ -102,14 +108,17 @@ public class BracketAction extends ScheduledAction {
 	}
 
 	public String bracketDelete() {
+		Authority auth = checkAuthority(buildResource(Modules.s_bracket_model, Operation.s_operation_delete));
+		if (auth != null) {
+			return auth.getName();
+		}
 		try {
 			m_bracket = m_bracketService.findByPK(m_bracketId);
 			m_documentService.deleteDocument(m_bracket.getDocumentId());
 			m_scheduleService.deleteSchedule(m_bracket.getScheduleId());
 			int count = m_bracketService.deleteBracket(m_bracketId);
 			if (count > 0) {
-				Log log = createLog(Constrants.s_bracket_model, Constrants.s_operation_delete,
-				      m_bracketId);
+				Log log = createLog(Modules.s_bracket_model, Operation.s_operation_delete, m_bracketId);
 
 				m_logService.insertLog(log);
 				return SUCCESS;
@@ -123,13 +132,16 @@ public class BracketAction extends ScheduledAction {
 	}
 
 	public String queryAllBrackets() {
-		m_brackets = m_bracketService.queryLimitedBrackets(m_tunnelId,
-		      m_tunnelSectionId, 0, Integer.MAX_VALUE);
+		m_brackets = m_bracketService.queryLimitedBrackets(m_tunnelId, m_tunnelSectionId, 0, Integer.MAX_VALUE);
 
 		return SUCCESS;
 	}
 
 	public String bracketList() {
+		Authority auth = checkAuthority(buildResource(Modules.s_bracket_model, Operation.s_operation_detail));
+		if (auth != null) {
+			return auth.getName();
+		}
 		try {
 			if (m_tunnelId == 0) {
 				m_tunnelId = m_tunnelService.queryDefaultTunnelId();
@@ -144,8 +156,7 @@ public class BracketAction extends ScheduledAction {
 			if (start < 0) {
 				start = 0;
 			}
-			m_brackets = m_bracketService.queryLimitedBrackets(m_tunnelId,
-			      m_tunnelSectionId, start, SIZE);
+			m_brackets = m_bracketService.queryLimitedBrackets(m_tunnelId, m_tunnelSectionId, start, SIZE);
 			for (Bracket bracket : m_brackets) {
 				bracket.setTunnel(m_tunnelService.findByPK(bracket.getTunnelId()));
 				int scheduleId = bracket.getScheduleId();
@@ -171,8 +182,8 @@ public class BracketAction extends ScheduledAction {
 		try {
 			m_bracket = m_bracketService.findByPK(m_bracketId);
 			m_tunnels = m_tunnelService.queryAllTunnels();
-			m_tunnelSections = m_tunnelSectionService.queryLimitedTunnelSectionsByTunnelId(
-			      m_bracket.getTunnelId(), 0, Integer.MAX_VALUE);
+			m_tunnelSections = m_tunnelSectionService.queryLimitedTunnelSectionsByTunnelId(m_bracket.getTunnelId(), 0,
+			      Integer.MAX_VALUE);
 			m_schedule = m_scheduleService.findByPK(m_bracket.getScheduleId());
 			m_constructionUnits = m_constructionUnitService.queryAllConstructionUnits();
 			int documentId = m_bracket.getDocumentId();
@@ -192,22 +203,25 @@ public class BracketAction extends ScheduledAction {
 	}
 
 	public String bracketUpdateSubmit() {
+		Authority auth = checkAuthority(buildResource(Modules.s_bracket_model, Operation.s_operation_update));
+		if (auth != null) {
+			return auth.getName();
+		}
 		try {
 			if (m_uploadFile.getFile() != null) {
 				int documentId = m_bracket.getDocumentId();
 				if (documentId > 0) {
 					Document document = m_documentService.findByPK(documentId);
-					m_documentService.updateDocument(Constrants.s_contactChannel_model, m_uploadFile, document);
+					m_documentService.updateDocument(Modules.s_bracket_model, m_uploadFile, document);
 				} else {
-					documentId = m_documentService.insertDocument(Constrants.s_contactChannel_model, m_uploadFile);
+					documentId = m_documentService.insertDocument(Modules.s_bracket_model, m_uploadFile);
 					m_bracket.setDocumentId(documentId);
 				}
 			}
 			m_scheduleService.updateSchedule(m_schedule);
 			int count = m_bracketService.updateBracket(m_bracket);
 			if (count > 0) {
-				Log log = createLog(Constrants.s_bracket_model, Constrants.s_operation_update,
-				      m_bracket);
+				Log log = createLog(Modules.s_bracket_model, Operation.s_operation_update, m_bracket);
 
 				m_logService.insertLog(log);
 				return SUCCESS;
